Fix Relation usage bug and improve memory usage
The memory usage is reduced for the typical applications of semimatroid
or orientability testing, which cache parts of their setup depending on
the cube they operate on. Many cubes lead to high memory usage, when
though most of the cubes in selfadhesivity testing are isomorphic.
This commits permutes the the relation instead so that we have only
one new cube per cardinality in the selfadhesivity testing.

There was one severe bug in the algorithm when it accessed the internals
of the CInet::Relation $S -- in the meantime the implementation details
changed.